Fix most liniting problems
