+++ src/main.py
@@ -1,5 +1,7 @@
-import numpy as np
-from pykalman import KalmanFilter
+"""Module which runs the navigation algorithm as a whole."""
+
+# import numpy as np
+# from pykalman import KalmanFilter
 
 
 def main():
@@ -16,9 +18,10 @@ def main():
 
         # Receiving data from vision
         vision = get_vision()
+        print(vision)
 
         # Update state given GPS data
-        current_state = update_state(gps)
+        current_state = update_state(current_state, gps)
 
         # Find optimal steering angle
         steering_angle = get_steering_angle(current_state)
@@ -36,28 +39,29 @@ def calculate_moving_average(state, data):
     """
     take average of current state and gps data
     """
-
-    return
+    print(state, data)
+    return 0
 
 
 def get_gps():
     """
     Receives data from GPS
     """
-    return
+    return 0
 
 
 def get_vision():
     """
     Receives data from vision
     """
-    return
+    return 0
 
 
 def update_state(state, gps):
     """
     Updates state given GPS data
     """
+    print(state, gps)
     return state
 
 
@@ -65,6 +69,7 @@ def get_steering_angle(state):
     """
     Returns optimal steering angle
     """
+    print(state)
     steering_angle = 0.0
 
     return steering_angle
@@ -74,7 +79,8 @@ def output_data(steering_angle):
     """
     Sends steering angle data to controls
     """
-    return
+    print(steering_angle)
+    return 0
 
 
 if __name__ == "__main__":
